+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
-
+## 2.6.0 (2021-04-26)
+### Changed
+- Decorate metrics with `mongodbClusterName`
+  `clusterName` (deprecated) attribute might get overwritten by the Agent when running inside kubernetes.
+  This change allows Mongodb entities to be uniquely identified when multiple Mongodb clusters are located in the same Kubernetes cluster.
+- removing obsolete win-config file
+- Use go 1.16 and go mod
+- Adds json schema integration integration tests  
 ## 2.5.1 (2021-02-09)
 ### Changed
 - `username` and `password` arguments are now optional